+++ abrechnung/database/revisions/0004_relax_db_constraints.sql
@@ -1,2 +1,2 @@
--- revision: f133b1d3
+-- migration: f133b1d3
 -- requires: b32893f6