+++ abrechnung/database/revisions/0009_robust_notifications.sql
@@ -1,4 +1,4 @@
--- revision: bf5fbd44
+-- migration: bf5fbd44
 -- requires: 156aef63
 
 alter table transaction_revision add column version int default 0;