+++ abrechnung/database/revisions/0013_register_with_invite.sql
@@ -1,4 +1,4 @@
--- revision: c019dd21
+-- migration: c019dd21
 -- requires: 5b333d87
 
 alter table usr add column is_guest_user bool default false not null;