+++ abrechnung/database/revisions/0015_metadata_fields_and_revision_changed.sql
@@ -1,4 +1,4 @@
--- revision: ee5d2b35
+-- migration: ee5d2b35
 -- requires: 174ef0fc
 
 alter table account_revision