+++ abrechnung/database/revisions/0017_clean_non_nullable_strings.sql
@@ -1,4 +1,4 @@
--- revision: 04424b59
+-- migration: 04424b59
 -- requires: a83f4798
 
 alter table transaction_history alter column description set default '';